refactor(ios): decompose IosPeripheral 713->390 loc via extension functions
